\name{GSE70531_params}
\alias{GSE70531_params}
\docType{data}
\title{
Parameters estimated for the MPRA data (GSE70531) 
}
\description{
Parameters estimated for the DNA and RNA (CONTROL) MPRA data (GSE70531) using the estimateMPRA function in this package. 
}
\usage{data("GSE70531_params")}
\format{
  The format is:
 num [1:231476] 1.86e-06 3.28e-06 3.22e-06 6.44e-06 2.03e-06 ...
}
\details{
This list contains the dispersion function across the input and output replicates estimated using DESeq2, the function to generate the proportion of input counts across tags,the function to generate the estimated transfection efficiencies across the oligos, and the normalization factor for the input and output replicates generated by DESeq2.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE70531}
}
\references{
Ulirsch JC, Nandakumar SK, Wang L, Giani FC et al. \emph{Systematic Functional Dissection of Common Genetic Variation Affecting Red Blood Cell Traits.} Cell 2016 Jun 2;165(6):1530-1545. PMID: 27259154
}
\examples{
data(GSE70531_params)
}
\keyword{datasets}
